Move the rest of generic diff functions to utils
