+++ CHANGES.rst
@@ -143,8 +143,8 @@ astropy.io.misc
 astropy.io.fits
 ^^^^^^^^^^^^^^^
 
-- ``report_diff_values()`` utility function is moved to ``astropy.utils.diff``.
-  [#7444]
+- ``diff_values()`, ``report_diff_values()``, and ``where_not_allclose()``
+  utility functions are moved to ``astropy.utils.diff``. [#7444]
 
 astropy.io.registry
 ^^^^^^^^^^^^^^^^^^^
@@ -191,8 +191,8 @@ astropy.utils
 ^^^^^^^^^^^^^
 
 - ``InheritDocstrings`` now also works on class properties. [#7166]
-- ``report_diff_values()`` utility function is moved from
-  ``astropy.io.fits.diff``. [#7444]
+- ``diff_values()`, ``report_diff_values()``, and ``where_not_allclose()``
+  utility functions are moved from ``astropy.io.fits.diff``. [#7444]
 
 astropy.visualization
 ^^^^^^^^^^^^^^^^^^^^^
