+++ astropy/utils/diff.py
@@ -8,10 +8,25 @@ import numpy as np
 
 from .misc import indent
 
-__all__ = ['fixed_width_indent', 'report_diff_values']
+__all__ = ['fixed_width_indent', 'diff_values', 'report_diff_values',
+           'where_not_allclose']
 
 # Smaller default shift-width for indent:
 fixed_width_indent = functools.partial(indent, width=2)
+
+
+def diff_values(a, b, rtol=0.0, atol=0.0):
+    """
+    Diff two scalar values.  If both values are floats they are compared to
+    within the given absolute and relative tolerance.
+    """
+
+    if isinstance(a, float) and isinstance(b, float):
+        if np.isnan(a) and np.isnan(b):
+            return False
+        return not np.allclose(a, b, rtol=rtol, atol=atol)
+    else:
+        return a != b
 
 
 def report_diff_values(fileobj, a, b, ind=0):
@@ -64,3 +79,22 @@ def report_diff_values(fileobj, a, b, ind=0):
                 line = ' ' * padding + line
         fileobj.write(
             fixed_width_indent('  {}\n'.format(line.rstrip('\n')), ind))
+
+
+def where_not_allclose(a, b, rtol=1e-5, atol=1e-8):
+    """
+    A version of numpy.allclose that returns the indices where the two arrays
+    differ, instead of just a boolean value.
+    """
+
+    # Create fixed mask arrays to handle INF and NaN; currently INF and NaN
+    # are handled as equivalent
+    if not np.all(np.isfinite(a)):
+        a = np.ma.fix_invalid(a).data
+    if not np.all(np.isfinite(b)):
+        b = np.ma.fix_invalid(b).data
+
+    if atol == 0.0 and rtol == 0.0:
+        # Use a faster comparison for the most simple (and common) case
+        return np.where(a != b)
+    return np.where(np.abs(a - b) > (atol + rtol * np.abs(b)))
